Add a few rpc integration test assertions, move test
